% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleep-package.R
\docType{package}
\name{acceleep-package}
\alias{acceleep}
\alias{acceleep-package}
\title{acceleep: Dummy Package}
\description{
This is only a dummy package to (hopefully) ease some workflows.
}
\author{
\strong{Maintainer}: Lukas Burk \email{lukas@quantenbrot.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})

}
\keyword{internal}
