% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-data.R
\name{aggregate_spiro}
\alias{aggregate_spiro}
\title{Take raw spirotmetry data and collapse it into approrpiate intervals}
\usage{
aggregate_spiro(input_file_spiro, t0, spiro_interval = 30)
}
\arguments{
\item{input_file_spiro}{File path to spirometry data.}

\item{t0}{Value of class \code{hms::hms()} to calibrate interval sequence to.}

\item{spiro_interval}{\verb{[30]} Interval in seconds to calculate mean METs from}
}
\value{
A tibble
}
\description{
Take raw spirotmetry data and collapse it into approrpiate intervals
}
\note{
Since t0 will vary ma acceloreter model, placement, and SID,
this function needs to be called for all combinations thereof.
}
\examples{
\dontrun{
aggregate_spiro("data/input/spiro/ID_001_spiro.csv", t0 = hms::as_hms("09:00:00"))
}
}
