% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{assemble_train_data}
\alias{assemble_train_data}
\title{Assemble training data for prediction}
\usage{
assemble_train_data(
  accel_model = c("actigraph", "activpal", "geneactiv"),
  placement = c("hip_left", "hip_right", "thigh_right", "wrist_left", "wrist_right"),
  outcome = c("MET", "kJ", "Jrel"),
  random_seed = 11235813,
  val_split = 1/3,
  interval_length = 30,
  res = 1
)
}
\arguments{
\item{placement}{One of \code{c("hip_left", "hip_right", "thigh_right", "wrist_left", "wrist_right")}.}

\item{outcome}{\verb{[MET]} one of \code{c("MET", "kJ", "Jrel")}, referring to
metabolic equivalents, absolute energy expenditure and relative energy
expenditure respectively.}

\item{random_seed}{\verb{[11235813]} Seed used to ensure reproducibility via \code{\link{set.seed}}.}

\item{val_split}{\verb{[1/3]} Validation split, meaning e.g. 1/3 of the subjects
(rounded down) will be used as validation dataset.}

\item{interval_length}{\verb{[30]} Number of seconds one interval covers.}

\item{res}{\verb{[100]}: Resolution of accelerometry data, in Hz. 20 for activPal,
100 for other models.}
}
\value{

}
\description{
Assemble training data for prediction
}
\section{Available Data}{


The following combinations of \code{model} and \code{placement} are available:\tabular{ll}{
   model \tab placement \cr
   actigraph \tab hip_left \cr
   actigraph \tab hip_right \cr
   activpal \tab thigh_right \cr
   geneactiv \tab hip_right \cr
   geneactiv \tab wrist_left \cr
   geneactiv \tab wrist_right \cr
}
}

