% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_clean_data}
\alias{combine_clean_data}
\title{Combine cleaned data for one accelerometer model / placement}
\usage{
combine_clean_data(
  model = c("actigraph", "activpal", "geneactiv"),
  placement = c("hip_left", "hip_right", "thigh_right", "wrist_left", "wrist_right")
)
}
\arguments{
\item{model}{Accelerometer model, one of \code{c("actigraph", "activpal", "geneactiv")}.}

\item{placement}{One of \code{c("hip_left", "hip_right", "thigh_right", "wrist_left", "wrist_right")}.}
}
\value{
A tibble. If combination of \code{model} and \code{placement} is invalid, an empty \code{data.frame}.
}
\description{
This function collects the datasets (split by subject) for a given model and placement, where the result is a
single dataset of all measurements across subjects.
}
\note{
Note that not every combination of \code{model} and \code{placement} are valid.
}
\examples{
\dontrun{
combine_clean_data("actigraph", "hip_left")
}
}
