% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{extract_outcome}
\alias{extract_outcome}
\title{Extract the outcome variables from the working dataset}
\usage{
extract_outcome(
  xdf,
  outcome = c("MET", "kJ", "Jrel"),
  output_type = c("numeric", "tbl"),
  rescale = FALSE
)
}
\arguments{
\item{xdf}{A full dataset containing outcome measures and interval / subject identifiers
as returned by \code{\link[=convert_input_data]{convert_input_data()}}  or combined by \code{\link[=combine_clean_data]{combine_clean_data()}}}

\item{outcome}{\verb{[MET]} one of \code{c("MET", "kJ", "Jrel")}, referring to
metabolic equivalents, absolute energy expenditure and relative energy
expenditure respectively.}

\item{output_type}{\verb{["numeric", "tbl"]} Return the outcome as a \code{"numeric"} vector
or as a \code{\link[tibble:tibble]{tibble::tibble()}}, where the latter additionally preserves
\code{interval} and \code{ID} variables for potential debugging.}

\item{rescale}{\verb{[FALSE]} If \code{TRUE}, outcome variable will be rescale using \code{\link[=scale]{scale()}}.}
}
\value{
Either a \code{numeric} vector (default) or \verb{[tibble::tibble()]}, depending on
\code{outcome_type}.
}
\description{
The working dataset will have both accelerometry and energy expenditure
measurements, where the latter are duplicated along each measurement interval.
This function extracts the outcome variable for each interval and returns it
so that the result is a vector with the same number of elements as there are
intervals in the dataset. Since intervals are numbered uniquely only per
subject, an additional variable \code{ID} is expected to ensure that datasets
containing multiple are handled correctly.
}
\examples{
\dontrun{
xdf <- combine_clean_data(model = "actigraph", placement = "hip_right")
mets <- extract_outcome(xdf, outcome = "MET", output_type = "numeric")

# mets is now a vector with one value per interval across all subjects
}
}
