% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_ts_chunk}
\alias{generate_ts_chunk}
\title{Generate a single interval of time-series of acceloremetry data}
\usage{
generate_ts_chunk(res = 20, interval_length = 30, offset = 0)
}
\arguments{
\item{res}{Resolution of accelormetry data, in Hz. 20 for Activpal, 100 for other models}

\item{interval_length}{Number of seconds a chunk covers}

\item{offset}{Second offset from previous chunk}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
One chunk of length \code{interval_length} seconds of accelerometry data
measured at \code{res} Hz with normally-distributed X/Y/Z measurements.
Using \code{offset} (in seconds) allows this function to be called repeatedly to
create consecutive chunks.
}
