% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_ts_dataset}
\alias{generate_ts_dataset}
\title{Generate a full acceloremtry dataset}
\usage{
generate_ts_dataset(res = 20, interval_length = 30, n_chunks = 100)
}
\arguments{
\item{res}{Resolution of accelerometry data, in Hz. 20 for Activpal, 100 for other models}

\item{interval_length}{Number of seconds a chunk covers}

\item{n_chunks}{Number of consecutive chunks to generate.}
}
\value{
a \link[tibble:tibble-package]{tibble} with \code{res * interval_length * n_chunks} rows.
}
\description{
Wraps \code{\link[=generate_ts_chunk]{generate_ts_chunk()}} to generate \code{n_chunks} consecutive chunks
of accelerometry data.
}
