% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-data.R
\name{get_demo}
\alias{get_demo}
\title{Get demographics}
\usage{
get_demo(file_demo = here::here("data/input/demo.csv"), id = NULL)
}
\arguments{
\item{file_demo}{\verb{[here::here("data/input/demo.csv")]}: Path to \code{demo.csv}.}

\item{id}{Optional. Specify a subject id such as \code{"005"} or \code{5} (will be coerced to numeric) to filter the demographics data
by that id.}
}
\value{
A tibble
}
\description{
Reads in the demographics data without further processing.
}
\examples{
\dontrun{
get_demo()

get_demo(id = "003")
}
}
