% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_overview_table}
\alias{get_overview_table}
\title{Create Table of Datasets}
\usage{
get_overview_table()
}
\value{
A tibble
}
\description{
Create Table of Datasets
}
\examples{
\dontrun{
get_overview_table()
}
}
