% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_from_path}
\alias{id_from_path}
\title{Extract the subject ID from a file path}
\usage{
id_from_path(x)
}
\arguments{
\item{x}{A file path, e.g. \code{"data/input/spiro/ID_001_spiro.csv"}}
}
\value{
A character of the same length as the input
}
\description{
This is entirely overengineered.
}
\examples{
id_from_path("data/input/spiro/ID_001_spiro.csv") # -> "001"
}
