% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfruns_analysis.R
\name{ingest_runs}
\alias{ingest_runs}
\title{Read tfruns with minor cleanup}
\usage{
ingest_runs(runs_dir)
}
\arguments{
\item{runs_dir}{Directory path used for \code{tfruns} to collect all the runs.}
}
\value{
A tibble
}
\description{
Read in all runs from a directory, calculate additionale the RMSE loss,
training duration in HMS and sort by ascending RMSE.
}
\examples{
\dontrun{
ingest_runs("output/runs/downsampled-ad-hoc/")
}
}
