% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{keras_prep_lstm}
\alias{keras_prep_lstm}
\alias{keras_prep_regression}
\title{Do the entire preparation dance}
\usage{
keras_prep_lstm(
  model = c("actigraph", "activpal", "geneactiv"),
  placement = c("hip_left", "hip_right", "thigh_right", "wrist_left", "wrist_right"),
  outcome = c("MET", "kJ", "Jrel"),
  random_seed = 11235813,
  val_split = 1/3,
  interval_length = 30,
  res = 100,
  normalize = TRUE
)

keras_prep_regression(
  model = c("actigraph", "activpal", "geneactiv"),
  placement = c("hip_left", "hip_right", "thigh_right", "wrist_left", "wrist_right"),
  outcome = c("MET", "kJ", "Jrel"),
  random_seed = 11235813,
  val_split = 1/3,
  interval_length = 30,
  res = 100,
  normalize = TRUE
)
}
\arguments{
\item{model}{Accelerometer model, one of \code{c("actigraph", "activpal", "geneactiv")}.}

\item{placement}{One of \code{c("hip_left", "hip_right", "thigh_right", "wrist_left", "wrist_right")}.}

\item{outcome}{\verb{[MET]} one of \code{c("MET", "kJ", "Jrel")}, referring to
metabolic equivalents, absolute energy expenditure and relative energy
expenditure respectively.}

\item{random_seed}{\verb{[11235813]} Seed used to ensure reproducibility via \code{\link{set.seed}}.}

\item{val_split}{\verb{[1/3]} Validation split, meaning e.g. 1/3 of the subjects
(rounded down) will be used as validation dataset.}

\item{interval_length}{\verb{[30]} Number of seconds one interval covers.}

\item{res}{\verb{[100]}: Resolution of accelerometry data, in Hz. 20 for activPal,
100 for other models.}
}
\value{
A nested list of the form\preformatted{list(
  training = list(
    data = train_data,
    labels = train_labels
  ),
  validation = list(
    data = test_data,
    labels = test_labels
  )
)
}

Where \code{train_data} and \code{test_data} are already converted to a tensor with 3
axes and \code{train_labels} and \code{test_labels} are numeric vectors of the same
length as the first axis of the respective \verb{_data} tensor.


}
\description{
This function wraps up all steps required to prepare the sample data
for \code{keras} modelling using RNNs (LSTM/GRU).
}
\details{
This is a wrapper around the following steps:
\enumerate{
\item \code{\link[=combine_clean_data]{combine_clean_data()}}: Aggregate per-subject datasets for a given
accelerometer model and placement.
\item \code{\link[=make_initial_splits]{make_initial_splits()}}: Split the data into a training and validation
dataset.
\item \code{\link[=split_data_labels]{split_data_labels()}}: Split the data into accelerometry and calorimetry
data.
\item \code{\link[=keras_reshape_accel]{keras_reshape_accel()}}: Reshape the accelerometry data into an array
(tensor) of shape (number_of_intervals, \code{res * interval_length}, 3).
}
}
\section{Available Data}{


The following combinations of \code{model} and \code{placement} are available:\tabular{ll}{
   model \tab placement \cr
   actigraph \tab hip_left \cr
   actigraph \tab hip_right \cr
   activpal \tab thigh_right \cr
   geneactiv \tab hip_right \cr
   geneactiv \tab wrist_left \cr
   geneactiv \tab wrist_right \cr
}
}

\examples{
\dontrun{
c(c(train_data, train_labels), c(test_data, test_labels)) \%<-\% keras_prep_lstm(
  model = "geneactiv", placement = "hip_right",
  outcome = "kJ", random_seed = 19283, val_split = 1/2,
  interval_length = 30, res = 1
)
}
}
