% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{keras_reshape_accel}
\alias{keras_reshape_accel}
\title{Convert a tbl of accelerometry to a Keras- / LSTM-friendly-array}
\usage{
keras_reshape_accel(accel_tbl, interval_length = 30, res = 100)
}
\arguments{
\item{accel_tbl}{Input tbl with \emph{only} columns X/Y/Z, one row per measurement.}

\item{interval_length}{\verb{[30]} Number of seconds one interval covers.}

\item{res}{\verb{[100]}: Resolution of accelerometry data, in Hz. 20 for activPal,
100 for other models.}
}
\value{
An \code{\link[=array]{array()}} of dimensions \code{c(samples, res * interval_length, 3)},
wherte \code{samples} is calculated as \code{nrow(accel_tbl) / (interval_length * res)}.
}
\description{
As this function uses \code{\link[keras:reexports]{keras::array_reshape()}}, its lack of sophistication
barely justifies its existence and one might as well just use
\code{\link[keras:reexports]{keras::array_reshape()}}.
}
\note{
This function uses \code{\link[keras:reexports]{keras::array_reshape()}}, which adds a rather daunting
dependency on Python, but at least it has saved the author from hours
of unsuccessful experimentation trying to re-create the correct output
structure using \code{\link[=array]{array()}}.
}
\examples{
\dontrun{
# Aggregating subject data for model/placement
full_data <- combine_clean_data("activpal", "thigh_right")

# Split into train / validation datasets
c(training_data, validation_data) \%<-\% make_initial_splits(
  full_data, random_seed = 21421, val_split = 1/3
)

# Split into data and labels
split_data <- split_data_labels(training_data, validation_data, outcome = "kJ")

c(train_data, train_labels) \%<-\% split_data$training
c(test_data, test_labels) \%<-\% split_data$validation

# Reshaping to array form
train_data <- keras_reshape_accel(train_data, 30, 20)
test_data <- keras_reshape_accel(test_data, 30, 20)
}
}
