% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-text.R
\name{label_accel_models}
\alias{label_accel_models}
\title{Label accelerometer models for final text}
\usage{
label_accel_models(x)
}
\arguments{
\item{x}{Name of accelerometer model in the dataset, e.g. \code{"actigraph"}}
}
\value{
A \code{character()} of the same length as the input
}
\description{
Label accelerometer models for final text
}
\examples{
label_accel_models("actigraph") # -> "Actigraph GT3X"
label_accel_models("activpal") # -> "activPAL"
label_accel_models("geneactiv") # -> "GENEActiv"
}
