% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-text.R
\name{label_outcome}
\alias{label_outcome}
\title{Label outcome units}
\usage{
label_outcome(x, with_abbr = FALSE)
}
\arguments{
\item{x}{A string correspnding to a unit as found in the dataset, e.g. \code{"kJ"}.}

\item{with_abbr}{\verb{[FALSE]} If \code{TRUE}, \code{"kJ"} and \code{"Jrel"} are appended "(AEE)"
and "(REE)" respectively.}
}
\value{
A character vector of the same length as \code{x}
}
\description{
Translates:
\itemize{
\item "kJ" -> "kJ/min"
\item "Jrel" -> "J/min/kg"
\item "MET" is kept as is.
}
}
\examples{
label_outcome("kJ")
label_outcome("Jrel")
label_outcome(c("kJ", "Jrel", "MET"))
label_outcome(c("kJ", "Jrel", "MET"), with_abbr = TRUE)
}
