% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-text.R
\name{label_placement}
\alias{label_placement}
\title{Label accelerometer placements}
\usage{
label_placement(x)
}
\arguments{
\item{x}{Placement, e.g. \code{"hip_right"}}
}
\value{
A \code{character()} of the same length as the input
}
\description{
Label accelerometer placements
}
\examples{
label_placement("hip_right") # -> "right hip"
label_placement("thigh_right") # -> "right thigh"
label_placement("wrist_left") # -> "left wrist"
}
