% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{make_initial_splits}
\alias{make_initial_splits}
\title{Split an entire dataset into training/validation data.
Please note that the resulting datasets are \emph{not} yet split into data and labels.}
\usage{
make_initial_splits(full_data, random_seed = 11235813, val_split = 1/3)
}
\arguments{
\item{full_data}{The full dataset as e.g. assempled via \code{\link[=combine_clean_data]{combine_clean_data()}}.}

\item{random_seed}{\verb{[11235813]} Seed used to ensure reproducibility via \code{\link{set.seed}}.}

\item{val_split}{\verb{[1/3]} Validation split, meaning e.g. 1/3 of the subjects
(rounded down) will be used as validation dataset.}
}
\value{
A list of two tibbles, each containing the full dataset incl.
accelerometry and calorimetry.
}
\description{
Split an entire dataset into training/validation data.
Please note that the resulting datasets are \emph{not} yet split into data and labels.
}
\examples{
\dontrun{
full_data <- combine_clean_data("activpal", "thigh_right")
c(training_data, validation_data) \%<-\% make_initial_splits(
  full_data, random_seed = 21421, val_split = 1/3
)
}
}
