% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-modelling.R
\name{metric_rmse}
\alias{metric_rmse}
\title{Custom metric wrapper for RMSE}
\usage{
metric_rmse(...)
}
\arguments{
\item{y_true, y_pred}{True and predicted values.}
}
\value{
An object of class \verb{"python.builtin.function" "python.builtin.object"}.
}
\description{
Custom metric wrapper for RMSE
}
\note{
See https://keras.rstudio.com/reference/metric_binary_accuracy.html
}
\examples{
\dontrun{
model \%>\% compile(
  optimizer = optimizer_rmsprop(),
  loss = "mse",
  metrics = metric_rmse
)
}
}
