% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{normalize_accelerometry}
\alias{normalize_accelerometry}
\title{Normalize accelerometry data}
\usage{
normalize_accelerometry(training_data, validation_data)
}
\arguments{
\item{training_data}{Datasets as created by e.g. \verb{[make_initial_splits()]}.}

\item{validation_data}{Datasets as created by e.g. \verb{[make_initial_splits()]}.}
}
\value{
A list of two tibbles, ordered same as input arguments.
}
\description{
Standard normalization by subtracting the mean and dividing by standard
deviation. Mean and SD are calculated from the entire training data (across
all subjects/intervals) and is used to standardize both training- and test-
sets with the same parameters.
}
\examples{
\dontrun{
normalize_accelerometry(training_data, validation_data)
}
}
