% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfruns_analysis.R
\name{plot_loss_history}
\alias{plot_loss_history}
\title{Plot a run's loss training history}
\usage{
plot_loss_history(run_dir)
}
\arguments{
\item{run_dir}{Path to a single run's data. Can also be a single row
of a data.frame as returned by \code{tfruns::ls_runs}}
}
\value{
A ggplot2 object
}
\description{
Plot a run's loss training history
}
\examples{
\dontrun{
plot_loss_history("output/runs/downsampled-ad-hoc/2020-07-24T14-48-54Z/")
}
}
