% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_on_training_set}
\alias{predict_on_training_set}
\title{Predict outcome for all data in the given training set}
\usage{
predict_on_training_set(model, training_data, interval_length = 30, res = 1)
}
\arguments{
\item{model}{A keras model to predict from.}

\item{interval_length}{\verb{[30]} Number of seconds one interval covers.}

\item{res}{\verb{[100]}: Resolution of accelerometry data, in Hz. 20 for activPal,
100 for other models.}
}
\value{
A tibble
}
\description{
Predict outcome for all data in the given training set
}
