% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{read_cv_results}
\alias{read_cv_results}
\title{Read CV results and bind them together}
\usage{
read_cv_results(path, latest_only = TRUE)
}
\arguments{
\item{path}{Path to CV results, e.g. \code{here::here("output/cross-validation/CNN")}.}

\item{latest_only}{\verb{[TRUE]} Only return the latest run results.}
}
\value{
A tibble
}
\description{
Read CV results and bind them together
}
\examples{
\dontrun{
# Iterate over all model output folders, read in latest CV runs, bind together
cv_results <- purrr::map_df(
 here::here("output/cross-validation", c("LM", "DNN", "CNN", "RNN")),
  read_cv_results
)
}
}
