% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{read_holdout_results}
\alias{read_holdout_results}
\title{Read holdout results and bind them together}
\usage{
read_holdout_results(path)
}
\arguments{
\item{path}{Path to holdout validation results, e.g. \code{here::here("output/holdout-validation/CNN")}.}
}
\value{
A tibble
}
\description{
Read holdout results and bind them together
}
\examples{
\dontrun{
# Iterate over all model output folders, read in latest CV runs, bind together
holdout_results <- purrr::map_df(
 here::here("output/holdout-validation", c("LM", "DNN", "CNN", "RNN")),
  read_holdout_results
)
}
}
