% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_output_data}
\alias{save_output_data}
\title{Save an R object to a local directory}
\usage{
save_output_data(x, dir = "")
}
\arguments{
\item{x}{An R object. The name of the object will be the filename, e.g.
"iris" -> \code{iris.rds}.}

\item{dir}{Optional: A subdirectory of \code{here::here("output")}}
}
\value{
Nothing
}
\description{
This is purely a convenience function for the lazy.
}
\examples{
\dontrun{
save_output_data(iris)
}
}
