% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{summarize_accelerometry}
\alias{summarize_accelerometry}
\title{Summarize accelerometry data}
\usage{
summarize_accelerometry(
  training_data,
  validation_data = NULL,
  normalize = TRUE
)
}
\arguments{
\item{training_data}{Datasets as created by e.g. \verb{[make_initial_splits()]}.}

\item{validation_data}{Datasets as created by e.g. \verb{[make_initial_splits()]}.}
}
\value{
A list of tibbles
}
\description{
Statistics calculated:
}
\details{
\itemize{
\item mean
\item SD
\item lag-1 autocorrelation
\item min, max
\item quantiles at 10, 25, 50, 75 and 90
}
}
\examples{
\dontrun{
summarize_accelerometry(training_data, validation_data)
}
}
