% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfruns_analysis.R
\name{tfrun_get_flags}
\alias{tfrun_get_flags}
\alias{tfrun_get_metrics}
\title{Get tfruns metadata}
\usage{
tfrun_get_flags(run_dir)

tfrun_get_metrics(run_dir)
}
\arguments{
\item{run_dir}{Path to a single run's data. Can also be a single row
of a data.frame as returned by \code{tfruns::ls_runs}}
}
\value{
A list for flags, tibble for metrics
}
\description{
Get tfruns metadata
}
