% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfruns_analysis.R
\name{tfrun_label_summary}
\alias{tfrun_label_summary}
\title{Assemble tfrun metata into distinct labels}
\usage{
tfrun_label_summary(run_dir)
}
\arguments{
\item{run_dir}{Path to a single run's data. Can also be a single row
of a data.frame as returned by \code{tfruns::ls_runs}}
}
\value{

}
\description{
Assemble tfrun metata into distinct labels
}
\examples{
\dontrun{
tfrun_label_summary("path/to/run")
}
}
