% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleep-package.R
\docType{package}
\name{acceleep-package}
\alias{acceleep}
\alias{acceleep-package}
\title{acceleep: A Packaged Analysis Project}
\description{
Analysis and model development project packaged to
    utilize package infrastructure for easier maintenance of uility
    functions.
}
\author{
\strong{Maintainer}: Lukas Burk \email{lukas@quantenbrot.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})

}
\keyword{internal}
