% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-data.R
\name{convert_input_data}
\alias{convert_input_data}
\title{Read accelerometry + spiro and save to disk in analysis-friendly state}
\usage{
convert_input_data(
  input_file_accel,
  input_file_spiro,
  file_demo = here::here("data/input/demo.csv"),
  ID,
  spiro_interval = 30,
  overwrite = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{input_file_accel, input_file_spiro}{File path to accelerometry/spirometrydata CSV.}

\item{file_demo}{\verb{[here::here("data/input/demo.csv")]}: Path to \code{demo.csv}.}

\item{ID}{Subject ID as character with 3 places, e.g. \code{"002"}.}

\item{spiro_interval}{\verb{[30]} Interval in seconds to calculate mean METs from}

\item{overwrite}{\verb{[TRUE]} Overwrite existing output files or leave them untouched?}

\item{verbose}{\verb{[FALSE]} If \code{TRUE}, the function signals input file and save location to the console.}
}
\value{
Invisibly + writtem data with colums \code{c("interval", "ID", "X", "Y", "Z", "MET", "kJ", "Jrel")}
}
\description{
Read accelerometry + spiro and save to disk in analysis-friendly state
}
\details{
Accelerometer model is currently detected from the file path in \code{input_file_accel}, which has to contain
the following strings to properly detect each model:
\itemize{
\item \code{"actigraph"}: GT3X ActiGraph
\item \code{"activpal"}: activPAL
\item \code{"geneactiv"}: GENEActiv
}
}
\examples{
\dontrun{
# actigraph
input_file_accel <- "data/input/actigraph/026_left_readable.csv"
# activpal
input_file_accel <- "data/input/activpal/026_ActivPal_readable.csv"
# geneactiv
input_file_accel <- "data/input/geneactiv/026_wrist_right_readable.csv"
# spiro
input_file_spiro <- "data/input/spiro/ID_026_spiro.csv"
ID <- "026"

convert_input_data(input_file_accel, input_file_spiro, ID = ID, overwrite = FALSE, verbose = TRUE)

}
}
