% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-data.R
\name{convert_tbl_array}
\alias{convert_tbl_array}
\title{Defunct because silly: Convert Accelerometer tbl to Array}
\usage{
convert_tbl_array(accel_tbl, interval_length, res)
}
\arguments{
\item{accel_tbl}{Input tbl with columns X/Y/Z, one row per measurement.}

\item{interval_length}{Number of seconds a chunk covers}

\item{res}{Resolution of accelerometry data, in Hz. 20 for Activpal, 100 for other models}
}
\value{
An array of dimensions c(chunks_per_tbl, 3, rows_per_chunk)
}
\description{
Defunct because silly: Convert Accelerometer tbl to Array
}
\examples{
\dontrun{
convert_tbl_array(accel_tbl, interval_length = 30, res = 20)
}
}
