% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-data.R
\name{keras_reshape_accel}
\alias{keras_reshape_accel}
\title{Convert a tbl of accelerometry to a Keras- / LSTM-friendly-array}
\usage{
keras_reshape_accel(accel_tbl, interval_length = 30, res = 100)
}
\arguments{
\item{accel_tbl}{Input tbl with \emph{only} columns X/Y/Z, one row per measurement.}

\item{interval_length}{Number of seconds a chunk covers}

\item{res}{Resolution of accelerometry data, in Hz. 20 for Activpal, 100 for other models}
}
\value{
An \code{\link[=array]{array()}} of dimensions \code{c(samples, res * interval_length, 3)},
wherte \code{samples} is calculated as \code{nrow(accel_tbl) / (interval_length * res)}.
}
\description{
As this function uses \code{\link[keras:reexports]{keras::array_reshape()}}, its lack of sophistication
barely justifies its existence and one might as well just use
\code{\link[keras:reexports]{keras::array_reshape()}}.
}
\note{
This function uses \code{\link[keras:reexports]{keras::array_reshape()}}, which adds a rather daunting
dependency on Python, but at least it has saved the author from hours
of unsuccessful experimentation trying to re-create the correct output
structure using \code{\link[=array]{array()}}.
}
\examples{
\dontrun{
keras_reshape_accel}
}
