% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{split_data_labels}
\alias{split_data_labels}
\title{Split training/validation datasets into separate data and labels}
\usage{
split_data_labels(
  training_data,
  validation_data,
  outcome = c("MET", "kJ", "Jrel")
)
}
\arguments{
\item{training_data, validation_data}{Datasets as created by e.g. \verb{[make_initial_splits()]}.}

\item{outcome}{\verb{["MET"]}: Outcome variable to be extracted, passed to \verb{[extract_outcome()]}.}
}
\value{
A nested list of the form\preformatted{list(
  training = list(
    data = train_data,
    labels = train_labels
  ),
  validation = list(
    data = test_data,
    labels = test_labels
  )
)
}
}
\description{
This function only splits accelerometry from energy expenditure.
}
\note{
As of now, this function only ever returns one label per interval of accelerometry data.
}
\examples{
\dontrun{
full_data <- combine_clean_data("activpal", "thigh_right")
c(training_data, validation_data) \%<-\% make_initial_splits(
  full_data, random_seed = 21421, val_split = 1/3
)
split_data <- split_data_labels(training_data, validation_data, outcome = "kJ")

c(train_data, train_labels) \%<-\% split_data$training
c(test_data, test_labels) \%<-\% split_data$validation
}
}
